Doctrine DBAL
- DBAL extension replaces sqlite extension
