+++ extensions/dbal/lib/Storage/Driver/Dbal/HistoryIterator.php
@@ -1,0 +1,98 @@
+<?php
+
+/*
+ * This file is part of the PHPBench package
+ *
+ * (c) Daniel Leech <daniel@dantleech.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace PhpBench\Extensions\Dbal\Storage\Driver\Dbal;
+
+use PhpBench\Storage\HistoryEntry;
+use PhpBench\Storage\HistoryIteratorInterface;
+
+/**
+ * Lazily load history entries from the database.
+ */
+class HistoryIterator implements HistoryIteratorInterface
+{
+    private $repository;
+    private $statement;
+    private $position = 0;
+    private $actualPosition = null;
+    private $current;
+
+    /**
+     * @param Repository $repository
+     */
+    public function __construct(Repository $repository)
+    {
+        $this->repository = $repository;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function current()
+    {
+        $this->init();
+        $current = $this->current;
+        $entry = new HistoryEntry(
+            $current['run_uuid'],
+            new \DateTime($current['run_date']),
+            $current['context'],
+            $current['vcs_branch']
+        );
+
+        return $entry;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function next()
+    {
+        $this->position++;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function key()
+    {
+        return $this->position;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function rewind()
+    {
+        $this->position = 0;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function valid()
+    {
+        $this->init();
+
+        return (bool) $this->current;
+    }
+
+    private function init()
+    {
+        if (null === $this->statement) {
+            $this->statement = $this->repository->getHistoryStatement();
+        }
+
+        if ($this->position !== $this->actualPosition) {
+            $this->current = $this->statement->fetch(\PDO::FETCH_ASSOC, \PDO::FETCH_ORI_ABS, $this->position);
+            $this->actualPosition = $this->position;
+        }
+    }
+}
