--- extensions/sqlite/benchmarks/phpbench.json
@@ -1,8 +1,0 @@
-{
-    "bootstrap": "../../../vendor/autoload.php",
-    "path": ".",
-    "extensions": [
-        "PhpBench\\Extensions\\Sqlite\\SqliteExtension"
-    ],
-    "storage": "sqlite"
-}
