+++ phpunit.xml.dist
@@ -13,8 +13,8 @@
         <testsuite name="PhpBench XDebug Extension">
             <directory>./extensions/xdebug/tests</directory>
         </testsuite>
-        <testsuite name="PhpBench Sqlite Extension">
-            <directory>./extensions/sqlite/tests</directory>
+        <testsuite name="PhpBench DBAL Extension">
+            <directory>./extensions/dbal/tests</directory>
         </testsuite>
     </testsuites>
